Add actions to DetailView
